requirements.txt
Flask
PyJWT
python-dotenv
Flask-CORS